Refactor: integrate mswms_demodata into mswms cli  (#2799)
