+++ tests/_test_mswms/test_mswms.py
@@ -50,6 +50,6 @@ def test_main():
     with mock.patch("mslib.mswms.mswms.argparse.ArgumentParser.parse_args",
                     return_value=argparse.Namespace(plot_types=None, version=False, update=False, gallery=False,
                                                     debug=False, logfile=None, action=None,
-                                                    host=None, port=None)):
+                                                    host=None, port=None, seed=False)):
         mswms.main()
     assert pytest_wrapped_e.typename == "SystemExit"
